z_interface_app/requirements.txt
# Web Framework
Flask>=2.2.5
Werkzeug>=2.0.0
gunicorn==20.1.0
flask-debugtoolbar==0.13.1

# Docker Support
docker>=6.0.1

# HTTP/Network
requests>=2.32.2
idna>=2.8

# Testing & Quality
pytest>=7.3.1
pytest-cov>=4.0.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
pycodestyle>=2.12.0

# Type Hints & Utils
typing-extensions>=4.11.0
python-dotenv>=0.19.2

# Monitoring
prometheus-client>=0.16.0

# Database (if using SQLAlchemy)
flask-sqlalchemy>=3.1.1
SQLAlchemy>=2.0.37

# Security Testing
safety>=3.2.14
safety-schemas>=0.0.10

# Performance Testing
locust>=2.32.9

# Browser Automation (if needed)
selenium>=4.28.1

# API Development (if needed)
fastapi>=0.115.6
pydantic>=2.9.2
pydantic-core>=2.23.4

# Code Formatting
autopep8>=2.3.2

# aiohttp
aiohttp>=3.8.1

# openai
openai>=0.10.0